Handled fractional counts in CountGrammar and subclasses.

+++ java/src/edu/ohsu/cslu/ella/StringCountGrammar.java
@@ -235,7 +235,7 @@ public final class StringCountGrammar implements CountGrammar {
     }
 
     @Override
-    public final int binaryRuleObservations(final String parent, final String leftChild, final String rightChild) {
+    public final float binaryRuleObservations(final String parent, final String leftChild, final String rightChild) {
 
         final HashMap<String, Object2IntMap<String>> leftChildMap = binaryRuleCounts.get(parent);
         if (leftChildMap == null) {
@@ -251,7 +251,7 @@ public final class StringCountGrammar implements CountGrammar {
     }
 
     @Override
-    public final int unaryRuleObservations(final String parent, final String child) {
+    public final float unaryRuleObservations(final String parent, final String child) {
 
         final Object2IntMap<String> childMap = unaryRuleCounts.get(parent);
         if (childMap == null) {
@@ -262,7 +262,7 @@ public final class StringCountGrammar implements CountGrammar {
     }
 
     @Override
-    public final int lexicalRuleObservations(final String parent, final String child) {
+    public final float lexicalRuleObservations(final String parent, final String child) {
 
         final Object2IntMap<String> childMap = lexicalRuleCounts.get(parent);
         if (childMap == null) {
@@ -436,7 +436,7 @@ public final class StringCountGrammar implements CountGrammar {
     }
 
     @Override
-    public final int observations(final String parent) {
+    public final float observations(final String parent) {
         int count = 0;
 
         if (binaryRuleCounts.containsKey(parent)) {
